Integrate material from Learn C++.
All of the things dropped are already covered directly in the main
lesson. A few of the sections in the retained material were dropped
due to not fitting at all in this course. The primarily kept material
covers comparing C++ to Python, which seems useful for those who need
an extra boost.

Nearly all the Learn C++ on pointers has been imported. The reason is
that it covers basic pointers in C++ in a way that is not done in the
main lesson. This could be streamlined better, but since the assignment
in the lesson is so easy, the additional reading possibly does not
impact the time required.

Week 1:
"Learn C++: Basics" is integrated as "C++ and Python Week 01"
"Learn C++: Computation" is dropped. Covered in lesson already.

Week 2:
"Learn C++: Class Syntax" is replaced by "C++ and Python Week 02".
"Learn C++: Method Definitions" is dropped. Covered, but the inline vs forward declare in the main lesson could be improved.
"Learn C++: Operators" is dropped. Covered in lesson already. (Optional)
"Learn C++: Inheritance" is dropped. Covered in lesson already. (Optional)

Week 3:
"Learn C++: Containers" is replaced by "C++ and Python Week 03".
"Learn C++: Templates" is dropped. Covered in lesson already. (Optional)

Week 4:
Learn C++ was not used. No change.

Week 5:
All Learn C++ lessons on pointers has been kept and converted.